requirements.txt
pandas
nltk
transformers
torch
markovify
requests
scikit-learn
Flask
google-generativeai
